+++ core/src/br/cefetmg/games/screens/MenuScreen.java
@@ -35,7 +35,6 @@ public class MenuScreen extends BaseScreen {
 
         // instancia a textura e a região de textura (usada para repetir)
         background = new TextureRegion(new Texture("menu-background.png"));
-
         // configura a textura para repetir caso ela ocupe menos espaço que o
         // espaço disponível
         background.getTexture().setWrap(
